Refactor blockchain tests in coreTest
