+++ core-test/src/test/scala/org/bitcoins/core/protocol/blockchain/BlockHeaderSpec.scala
@@ -1,13 +1,15 @@
 package org.bitcoins.core.protocol.blockchain
 
 import org.bitcoins.testkitcore.gen.BlockchainElementsGenerator
-import org.scalacheck.{Prop, Properties}
+import org.bitcoins.testkitcore.util.BitcoinSUnitTest
 
 /** Created by tom on 7/6/16.
   */
-class BlockHeaderSpec extends Properties("BlockHeaderSpec") {
-  property("serialization symmetry") =
-    Prop.forAll(BlockchainElementsGenerator.blockHeader) { header =>
-      BlockHeader(header.hex) == header
+class BlockHeaderSpec extends BitcoinSUnitTest {
+
+  it must "serialization symmetry" in {
+    forAll(BlockchainElementsGenerator.blockHeader) { header =>
+      assert(BlockHeader(header.hex) == header)
     }
+  }
 }
